requirements.txt
numpy
transformers
soundfile
librosa
einops
vector_quantize_pytorch
tqdm
loguru
click
gradio